fix: Allow to re-run the script, again

+++ install_openvpn.sh
@@ -120,7 +120,9 @@ if [ -f /etc/nginx/sites-enabled/default ]; then
     sudo rm /etc/nginx/sites-enabled/default
 fi
 sudo cp etc/nginx/sites-available/easy_openvpn /etc/nginx/sites-available/
-sudo ln -s /etc/nginx/sites-available/easy_openvpn /etc/nginx/sites-enabled/
+if [ ! -e /etc/nginx/sites-enabled/easy_openvpn ]; then
+    sudo ln -s /etc/nginx/sites-available/easy_openvpn /etc/nginx/sites-enabled/
+fi
 sudo cp etc/systemd/system/easy_openvpn.service /etc/systemd/system/
 
 # Networking foo
